Created server and client packages and moved files there.
